+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
 - Added new version of `Hubspot` connector and test files.
+- Added `Mindful` connector and test file.
+
 
 ### Added